Bring wordpress_xmlrpc_login back, make wordpress_multicall as new
